+++ server/komodo-server/src/main/java/org/komodo/datasources/PostgreSQLDefinition.java
@@ -62,6 +62,7 @@ public class PostgreSQLDefinition extends DataSourceDefinition {
         props.setProperty("username", source.getProperty("user"));
         props.setProperty("password", source.getProperty("password"));
         props.setProperty("schema", source.getProperty("schema"));
+        props.setProperty("syndesis-connector-id", source.getId());
         return props;
     }
